Add /schema endpoint to C++ services and use it instead of /pack_action in tests
