+++ doc/psidk/src/default-apps/events.md
@@ -1,4 +1,3 @@
 # events
 
-{{#cpp-doc ::UserService::ServiceSchema}}
 {{#cpp-doc ::UserService::EventIndex}}
