+++ libraries/psio/src/schema.cpp
@@ -1354,13 +1354,10 @@ namespace psio::schema_types
       }
       for (auto& [name, type] : schema.types)
       {
-         if (auto* alias = std::get_if<Type>(&type.value))
-         {
-            auto& bestName = renamer.fixedNames[renamer.getAlias(alias->type)].bestName;
-            if (bestName == "@" || compareAlias(name, bestName) < 0)
-               bestName = name;
-         }
-         else
+         auto& bestName = renamer.fixedNames[renamer.getAlias(name)].bestName;
+         if (bestName == "@" || compareAlias(name, bestName) < 0)
+            bestName = name;
+         if (!std::holds_alternative<Type>(type.value))
          {
             visitTypes(renamer.addRef(), type);
          }
