+++ services/system/Producers/src/RProducers.cpp
@@ -61,6 +61,9 @@ std::optional<HttpReply> RProducers::serveSys(HttpRequest request)
    if (auto result = servePackAction<Producers>(request))
       return result;
 
+   if (auto result = serveSchema<Producers>(request))
+      return result;
+
    if (auto result = serveContent(request, Tables{getReceiver()}))
       return result;
