+++ services/user/Invite/src/AuthInvite.cpp
@@ -77,6 +77,9 @@ namespace UserService
       if (auto result = psibase::servePackAction<AuthInvite>(request))
          return result;
 
+      if (auto result = psibase::serveSchema<AuthInvite>(request))
+         return result;
+
       if (auto result = psibase::serveContent(request, Tables{}))
          return result;
